requirements.txt
pyautogui
ffmpeg-python
